api/requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
fullGSapi==1.3.11
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.0
gspread
google-auth
pydantic==2.9.2
psycopg2
